# Optimize ipset invoking in ipvs proxy