@lbernail 2020-09-03
I remember seeing this when I looked at flamegraphs in the latest optimization we did
This is great news
Can you edit the PR to explain the logic change (reduce calls to ipset by grouping changes if I understand the patch correctly)

Also I checked again and the netlink library still doesn't support ipsets (this would be a lot nicer...)