@Lion-Wei 2020-09-03
> Can you edit the PR to explain the logic change (reduce calls to ipset by grouping changes if I understand the patch correctly)

Sure thing, I added another description in this pr.
```
Reduce ipset call to only 1 ipset save and 1 ipset restore call in each sync.
Reorganized ipset related logic, add ipsetManager struct to manager ipsets entries cache, make more efficient and easier to read.
```

> Also I checked again and the netlink library still doesn't support ipsets (this would be a lot nicer...)

Yes, I also checked it before start this pr.