@Lion-Wei 2020-09-03
BTW, I noticed bind and unbind addresses from `kube-ipvs0` also takes lot time in one sync, and bind address to `kube-ipvs0` will result in many other issues. 

Recently I was trying to solve this problem, luckly already got some progress, I'll start an issue to discuss this.