@Lion-Wei 2020-09-22
/test pull-kubernetes-bazel-test